src/app/requirements.txt
# Web framework and ASGI server
fastapi
uvicorn[standard]

# Database and ORM
sqlalchemy
psycopg2-binary
alembic

# Auth and JWT
authlib
python-jose[cryptography]

# Environment variables
python-dotenv

# AWS S3 / MinIO
boto3
botocore

# Data validation
pydantic

# Browser automation
playwright
# HTTP client
httpx

# Security
itsdangerous

# Form parsing (for FastAPI file uploads)
python-multipart

# Testing
pytest

google-auth
google-auth-oauthlib
google-api-python-client

# Video/Audio processing
ffmpeg-python

# OpenAI API
openai
